Adding templates and view modules
Navigation, search, and results JS functionality implemented. We capture
an input string, obfuscate it, and then trigger an event passing the
string. This event is picked up in the app module which navigates to the
'results' route passing the string. The results view module then
identifies which images are available and renders them to the DOM.